requirements.txt
absl-py==0.13.0
altgraph==0.17.2
amqp==5.1.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
attrs==21.4.0
Authlib==1.0.1
auto-py-to-exe==2.8.0
autoit==0.2.4
beautifulsoup4==4.11.1
billiard==3.6.4.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.2
cachetools==4.2.2
Camelot==12.6.29
celery==5.2.7
certifi==2021.5.30
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.3
chromedriver-autoinstaller==0.3.1
click==8.1.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
comtypes==1.1.11
configparser==5.1.0
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==3.3.1
cx-Oracle==8.2.0
cycler==0.11.0
decorator==5.0.9
defusedxml==0.7.1
distro==1.5.0
Django==3.2.5
djangorestframework==3.12.4
dnspython==2.2.1
docutils==0.18.1
docx-mailmerge==0.5.0
Eel==0.12.4
Elixir==0.7.1
et-xmlfile==1.0.1
exchangelib==4.7.3
fire==0.4.0
flatbuffers==1.12
fonttools==4.34.4
fpdf2==2.5.5
furl==2.1.3
future==0.18.2
gast==0.4.0
gevent==21.1.2
gevent-websocket==0.10.1
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.13.2
greenlet==1.0.0
grpcio==1.34.1
h5py==3.1.0
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.2
install==1.3.4
isodate==0.6.1
itypes==1.2.0
java-access-bridge-wrapper==0.9.4
Jinja2==2.11.2
jsonpath-ng==1.5.3
jsonschema==4.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
kombu==5.2.4
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.5.2
MouseInfo==0.1.3
mss==6.1.0
mysql-connector-python==8.0.25
netsuitesdk==1.24.0
notifiers==1.3.3
ntlm-auth==1.5.0
numpy==1.19.5
oauthlib==3.1.1
openapi-codec==1.3.2
opencv-python==4.5.5.62
openpyxl==3.0.10
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.3
pandas==1.3.0
pbr==5.6.0
pdfminer==20191125
pdfminer.six==20201018
pdfminer3k==1.3.4
pefile==2021.9.3
Pillow==9.1.1
platformdirs==2.5.2
ply==3.11
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyAutoIt==0.6.5
pycparser==2.20
pycryptodome==3.10.1
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2021.3
PyJWT==1.7.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput-robocorp-fork==4.0.0
pyparsing==3.0.9
PyPDF2==1.28.4
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
pytweening==1.0.4
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==5.4.1
redis==3.5.3
requests==2.28.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
robocorp-dialog==0.5.3
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.3
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==5.1.3
robotframework-seleniumtestability==1.1.0
rpa==1.47.0
rpaframework==15.1.4
rpaframework-core==9.0.1
rpaframework-dialogs==3.0.0
rpaframework-pdf==4.0.2
rpaframework-windows==5.0.1
rsa==4.7.2
selenium==3.141.0
simple-salesforce==1.11.6
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.1
tabula-py==2.2.0
tabulate==0.8.9
tagui==1.47.0
Tempita==0.5.2
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tqdm==4.56.0
tweepy==3.10.0
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==2.1
uiautomation==2.0.16
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.2
webdrivermanager==0.10.0
Werkzeug==2.0.1
whichcraft==0.6.1
wrapt==1.12.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.1.0
zope.event==4.5.0
zope.interface==5.4.0
